scraper/requirements.txt
openai==1.3.0
langchain==0.0.350
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
lxml==4.9.3
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
redis==5.0.1
psycopg2-binary==2.9.9 